+++ data_spec_validator/spec/custom_spec/__init__.py
@@ -1,1 +1,3 @@
 from .defines import register
+
+__all__ = ["register"]